feat: Implement Discord voice commands and transcript handling
- Added local Discord voice stack setup instructions in README.md.
- Updated .env.example to include new environment variables for voice services.
- Enhanced docker-compose.yml to integrate voip_service with Discord service.
- Created a new Alembic migration to enforce uniqueness in platform_routes for voice.
- Modified PlatformRoute model to include container_id for voice routing.
- Implemented voice route management in TalkService, allowing for upsert and retrieval of voice routes.
- Enhanced VoiceTranscriptPublisher to handle coalescing and posting of transcripts with rate limit handling.
- Added integration tests for voice transcript functionality and unit tests for voice route management.
- Updated tasks.md to reflect completion of voice binding persistence and related features.

+++ services/memory_service/src/api/memories_service.py
@@ -220,12 +220,10 @@ class MemoriesService:
         ):
             offset_s = max(
                 0,
-                int(
-                    math.floor(
-                        (
-                            _ensure_utc(view.timestamp) - _ensure_utc(view.source.published_at)
-                        ).total_seconds()
-                    )
+                math.floor(
+                    (
+                        _ensure_utc(view.timestamp) - _ensure_utc(view.source.published_at)
+                    ).total_seconds()
                 ),
             )
             separator = "&" if "?" in view.source.url else "?"
